requirements.txt
Django==1.11.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3
scikit-learn==0.18.1
requests==2.11.1
gunicorn==19.7.1